Apply Phase 1 immediate security fixes (UserFront)

+++ UserFront/src/main/java/com/userFront/domain/PrimaryTransaction.java
@@ -20,7 +20,7 @@ public class PrimaryTransaction {
 	private String description;
 	private String type;
 	private String status;
-	private double amount;
+	private BigDecimal amount;
 	private BigDecimal availableBalance;
 	
 	@ManyToOne
@@ -31,7 +31,7 @@ public class PrimaryTransaction {
 
 	}
 
-	public PrimaryTransaction(Date date, String description, String type, String status, double amount,
+	public PrimaryTransaction(Date date, String description, String type, String status, BigDecimal amount,
 			BigDecimal availableBalance, PrimaryAccount primaryAccount) {
 		this.date = date;
 		this.description = description;
@@ -82,11 +82,11 @@ public class PrimaryTransaction {
 		this.status = status;
 	}
 
-	public double getAmount() {
+	public BigDecimal getAmount() {
 		return amount;
 	}
 
-	public void setAmount(double amount) {
+	public void setAmount(BigDecimal amount) {
 		this.amount = amount;
 	}
