+++ UserFront/src/main/java/com/userFront/service/AccountService.java
@@ -1,5 +1,6 @@
 package com.userFront.service;
 
+import java.math.BigDecimal;
 import java.security.Principal;
 
 import com.userFront.domain.PrimaryAccount;
@@ -10,6 +11,6 @@ public interface AccountService {
 	PrimaryAccount createPrimaryAccount();
 	SavingsAccount createSavingsAccount();
 	
-	void deposit(String accountType, double amount, Principal principal);
-	void withdraw(String accountType, double amount, Principal principal);
+	void deposit(String accountType, BigDecimal amount, Principal principal);
+	void withdraw(String accountType, BigDecimal amount, Principal principal);
 }
